Change name of template xml file to better represent the vx metrics it's used for.
Add the verifiction mask (vx_mask) as a property of databases and as an argument to make_single_mv_vx_plot.py.

Add aotk, pm25, and pm10 as valid forecast fields to validate.

Add the boolean "fcst_is_ensemble" as a property of METviewer databases.

Add fbar, obar, fbar_obar, rmse, bcrmse, and gss as valid vx metrics to plot.

Changes that should have been included with previous commit:
1) Initiale jinja2 variables in correct locations.
2) Define plot line type, width, and color for obs.

For convenience, introduce jinja2 variables boot_replics and boot_rand_seed for the number of replications and the random seed when performing aggregation statistics with bootstrapping.

Add 8m (8 meters) as a valid level.  This is needed because pm25 and pm10 obs are at this level.  This change should have been part of commit hash 4c2510fbc.

Remove extra space.

Add flag that determines whether or not the base rate is included in the plot.

Clean up comments.

Rename variable for clarity; clean up comments.

Add key to valid forecast fields that specifies whether the field is continuous (or discontinuous).

Add new METviewer database.

Add key to each valid vx metric that specifies whether or not we can plot confidence intervals (CIs) in METviewer for that metric.

Bug fix.

Introduce in jinja2 template a flag that determines whether or not confidence intervals are plotted (plot_conf_intervals); always use boot strapping for statistics (instead of using summary statistics).

Make flag for plotting confidence intervals a command-line argument to make_single_mv_vx_plot.py (with a default of True unless plotting CIs is not possible for the given metric).  This is instead of hard-coding it in the jinja2 template.

Add a multi-plot configuration parameter (plot_CIs_all_metrics) that specifies whether or not to (try to) plot confidence intervals (CIs) in each of the vx metric plots to be generated.

Bug fix: Add the vx mask to use to the config file for multi-plot generation.  This should have been added as part of commit 34739b617.

Remove debugging code.

Clean up unused lines.

Modify argument parser so that default values of arguments are printed when the help is invoked from the command line.

Add a multi-plotting configuration file for the database "mv_rrfsv1_science_eval" that shows how to generate vx plots for AOTK, PM10, and PM25.

Disable plotting of base rate; clean up comments.